requirements.txt
pyyaml
boto3
pyboto3
expiringdict
sanic
http-basic-auth
requests
python-jose